+++ CHANGELOG.md
@@ -5,6 +5,9 @@ See [Keep A Changelog](https://github.com/olivierlacan/keep-a-changelog)
 before changing this file!
 
 ## [Unreleased][unreleased]
+### Added
+- added method `VisMon.publish`
+
 ## [0.6.1] - 2015-03-17
 ### Fixed
 - prevent multiple executions of a throttled function